+++ src/flows/analyseImageFlow.ts
@@ -6,10 +6,9 @@ import { BaileysProvider as Provider } from "@builderbot/provider-baileys";
 import fs from "fs/promises";
 import { typing } from "../utils/presence";
 import sharp from "sharp";
-import { callOllamaAPI } from "./welcomeFlow.flow";
 import { createMessageQueue, QueueConfig } from "../utils/fast-entires";
-import { Messages } from "openai/resources/beta/threads/messages";
-import { Session, sessions } from "./welcomeFlow.flow";
+import { Session, sessions } from "../models/Session";
+import { callOllamaAPI } from "../services/ollamaService";
 
 const queueConfig: QueueConfig = { gapSeconds: 0 };
 const enqueueMessage = createMessageQueue(queueConfig);
